requirements.txt
# SRE Incident Response Environment

# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
websockets>=12.0

# Utilities
numpy>=1.24.0

# LLM Inference
openai>=1.0.0
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.25.0
